chore: production readiness P0/P1 improvements
- Remove committed .env; add .env.example and .gitignore rules

- Split router.rs and app.rs into focused submodules

- Add audit workflow, LFS gitattributes, README badges, branch prune script
