+++ .env.example
@@ -1,5 +1,5 @@
-# Placeholder environment variables for local documentation/examples
-# Replace with real values before running database-backed samples.
+# Placeholder environment variables for local documentation/examples.
+# Copy to .env and replace with real values before running database-backed samples.
 DATABASE_URL=postgres://postgres:postgres@localhost:5432/rustapi_dev
 REDIS_URL=redis://127.0.0.1:6379
 OAUTH_CLIENT_ID=replace-me
