+++ .gitattributes
@@ -1,4 +1,3 @@
-# Auto detect text files and perform LF normalization
-* text=auto
-api/public/*.txt text eol=lf
-.github/scripts/*.sh text eol=lf
+# Large binary assets — use Git LFS for PNGs over 500 KB
+assets/*.png filter=lfs diff=lfs merge=lfs -text
+assets/**/*.png filter=lfs diff=lfs merge=lfs -text
