+++ crates/rustapi-core/src/app/builder.rs
@@ -1,174 +1,27 @@
-//! RustApi application builder
-
+use super::config::RustApiConfig;
+use super::dispatcher::RequestDispatcher;
+use super::helpers::{add_path_params_to_operation, normalize_prefix_for_openapi};
+#[cfg(feature = "swagger-ui")]
+use super::helpers::{check_basic_auth, unauthorized_response};
+#[cfg(feature = "dashboard")]
+use super::helpers::{
+    infer_route_feature_gates, is_dashboard_replay_eligible, openapi_tags_for_route,
+};
+use super::production::ProductionDefaultsConfig;
+use super::types::RustApi;
 use crate::error::Result;
 use crate::events::LifecycleHooks;
 use crate::interceptor::{InterceptorChain, RequestInterceptor, ResponseInterceptor};
-use crate::middleware::{
-    BodyLimitLayer, BoxedNext, LayerStack, MiddlewareLayer, DEFAULT_BODY_LIMIT,
-};
+use crate::middleware::{BodyLimitLayer, LayerStack, MiddlewareLayer, DEFAULT_BODY_LIMIT};
 use crate::response::IntoResponse;
 use crate::router::{MethodRouter, Router};
 use crate::server::Server;
-use crate::{Request, Response};
-use http::Extensions;
 use std::collections::BTreeMap;
 #[cfg(feature = "dashboard")]
 use std::collections::BTreeSet;
 use std::future::Future;
 use std::sync::Arc;
 use tracing_subscriber::{layer::SubscriberExt, util::SubscriberInitExt, EnvFilter};
-
-/// A dispatcher that can drive requests through the RustAPI pipeline
-/// (interceptors + layers + router) without any network or serialization overhead.
-///
-/// Obtained via [`RustApi::request_dispatcher`].
-#[derive(Clone)]
-pub struct RequestDispatcher {
-    router: Arc<Router>,
-    layers: LayerStack,
-    interceptors: InterceptorChain,
-}
-
-impl RequestDispatcher {
-    /// Returns the shared state Extensions from the underlying router.
-    /// Useful for in-process request construction to preserve `State<T>` etc.
-    pub fn state_ref(&self) -> Arc<Extensions> {
-        self.router.state_ref()
-    }
-
-    /// Dispatch a request through the full stack (interceptors → middleware layers
-    /// → route handler → response interceptors).
-    ///
-    /// This replicates the logic used by the normal HTTP server.
-    pub async fn dispatch(&self, request: Request) -> Response {
-        let req = self.interceptors.intercept_request(request);
-
-        let path = req.path().to_owned();
-        let method = req.method().clone();
-
-        let response = if self.layers.is_empty() {
-            crate::server::route_request_direct(&self.router, req, &path, &method).await
-        } else {
-            let router = self.router.clone();
-            let p = path.clone();
-            let m = method.clone();
-
-            let routing_handler: BoxedNext = Arc::new(move |r: Request| {
-                let router = router.clone();
-                let pp = p.clone();
-                let mm = m.clone();
-                Box::pin(async move { crate::server::route_request(&router, r, &pp, &mm).await })
-            });
-
-            self.layers.execute(req, routing_handler).await
-        };
-
-        self.interceptors.intercept_response(response)
-    }
-}
-
-/// Main application builder for RustAPI
-///
-/// # Example
-///
-/// ```rust,ignore
-/// use rustapi_rs::prelude::*;
-///
-/// #[tokio::main]
-/// async fn main() -> Result<()> {
-///     RustApi::new()
-///         .state(AppState::new())
-///         .route("/", get(hello))
-///         .route("/users/{id}", get(get_user))
-///         .run("127.0.0.1:8080")
-///         .await
-/// }
-/// ```
-pub struct RustApi {
-    router: Router,
-    openapi_spec: rustapi_openapi::OpenApiSpec,
-    layers: LayerStack,
-    body_limit: Option<usize>,
-    interceptors: InterceptorChain,
-    lifecycle_hooks: LifecycleHooks,
-    hot_reload: bool,
-    #[cfg(feature = "http3")]
-    http3_config: Option<crate::http3::Http3Config>,
-    health_check: Option<crate::health::HealthCheck>,
-    health_endpoint_config: Option<crate::health::HealthEndpointConfig>,
-    status_config: Option<crate::status::StatusConfig>,
-    #[cfg(feature = "dashboard")]
-    dashboard_config: Option<crate::dashboard::DashboardConfig>,
-}
-
-/// Configuration for RustAPI's built-in production baseline preset.
-///
-/// This preset bundles together the most common foundation pieces for a
-/// production HTTP service:
-/// - request IDs on every response
-/// - structured tracing spans with service metadata
-/// - standard `/health`, `/ready`, and `/live` probes
-#[derive(Debug, Clone)]
-pub struct ProductionDefaultsConfig {
-    service_name: String,
-    version: Option<String>,
-    tracing_level: tracing::Level,
-    health_endpoint_config: Option<crate::health::HealthEndpointConfig>,
-    enable_request_id: bool,
-    enable_tracing: bool,
-    enable_health_endpoints: bool,
-}
-
-impl ProductionDefaultsConfig {
-    /// Create a new production baseline configuration.
-    pub fn new(service_name: impl Into<String>) -> Self {
-        Self {
-            service_name: service_name.into(),
-            version: None,
-            tracing_level: tracing::Level::INFO,
-            health_endpoint_config: None,
-            enable_request_id: true,
-            enable_tracing: true,
-            enable_health_endpoints: true,
-        }
-    }
-
-    /// Annotate tracing spans and default health payloads with an application version.
-    pub fn version(mut self, version: impl Into<String>) -> Self {
-        self.version = Some(version.into());
-        self
-    }
-
-    /// Set the tracing log level used by the preset tracing layer.
-    pub fn tracing_level(mut self, level: tracing::Level) -> Self {
-        self.tracing_level = level;
-        self
-    }
-
-    /// Override the default health endpoint paths.
-    pub fn health_endpoint_config(mut self, config: crate::health::HealthEndpointConfig) -> Self {
-        self.health_endpoint_config = Some(config);
-        self
-    }
-
-    /// Enable or disable request ID propagation.
-    pub fn request_id(mut self, enabled: bool) -> Self {
-        self.enable_request_id = enabled;
-        self
-    }
-
-    /// Enable or disable structured tracing middleware.
-    pub fn tracing(mut self, enabled: bool) -> Self {
-        self.enable_tracing = enabled;
-        self
-    }
-
-    /// Enable or disable built-in health endpoints.
-    pub fn health_endpoints(mut self, enabled: bool) -> Self {
-        self.enable_health_endpoints = enabled;
-        self
-    }
-}
 
 impl RustApi {
     /// Create a new RustAPI application
@@ -208,7 +61,7 @@ impl RustApi {
     /// The primary way to build a RustAPI application.
     ///
     /// Collects all routes decorated with `#[rustapi_rs::get]`, `#[rustapi_rs::post]`, etc.
-    /// at link time via `linkme` and registers them automatically — no manual `.route()`
+    /// at link time via `linkme` and registers them automatically â€” no manual `.route()`
     /// or `.mount_route()` calls needed. This is baked into the core and requires no
     /// feature flags.
     ///
@@ -449,7 +302,7 @@ impl RustApi {
     /// ```rust,ignore
     /// RustApi::new()
     ///     .on_start(|| async {
-    ///         println!("🚀 Server starting...");
+    ///         println!("ğŸš€ Server starting...");
     ///         // e.g. run DB migrations, warm caches
     ///     })
     ///     .run("127.0.0.1:8080")
@@ -476,7 +329,7 @@ impl RustApi {
     /// ```rust,ignore
     /// RustApi::new()
     ///     .on_shutdown(|| async {
-    ///         println!("👋 Server shutting down...");
+    ///         println!("ğŸ‘‹ Server shutting down...");
     ///         // e.g. flush logs, close DB connections
     ///     })
     ///     .run_with_shutdown("127.0.0.1:8080", ctrl_c())
@@ -565,7 +418,7 @@ impl RustApi {
         }
     }
 
-    fn mount_auto_routes_grouped(mut self) -> Self {
+    pub(super) fn mount_auto_routes_grouped(mut self) -> Self {
         let routes = crate::auto_route::collect_auto_routes();
 
         if routes.is_empty() {
@@ -1304,10 +1157,10 @@ impl RustApi {
             .map(|v| v == "1")
             .unwrap_or(false);
 
-        tracing::info!("🔄 Hot-reload mode enabled");
+        tracing::info!("ğŸ”„ Hot-reload mode enabled");
 
         if is_under_watcher {
-            tracing::info!("   File watcher active — changes will trigger rebuild + restart");
+            tracing::info!("   File watcher active â€” changes will trigger rebuild + restart");
         } else {
             tracing::info!("   Tip: Run with `cargo rustapi run --watch` for automatic hot-reload");
         }
@@ -1823,1147 +1676,8 @@ impl RustApi {
     }
 }
 
-#[cfg(feature = "dashboard")]
-fn openapi_tags_for_route(
-    spec: &rustapi_openapi::OpenApiSpec,
-    path: &str,
-    methods: &[http::Method],
-) -> Vec<String> {
-    let Some(path_item) = spec.paths.get(path) else {
-        return Vec::new();
-    };
-
-    let mut tags = BTreeSet::new();
-    for method in methods {
-        if let Some(operation) = operation_for_method(path_item, method) {
-            tags.extend(operation.tags.iter().cloned());
-        }
-    }
-
-    tags.into_iter().collect()
-}
-
-#[cfg(feature = "dashboard")]
-fn operation_for_method<'a>(
-    path_item: &'a rustapi_openapi::PathItem,
-    method: &http::Method,
-) -> Option<&'a rustapi_openapi::Operation> {
-    match *method {
-        http::Method::GET => path_item.get.as_ref(),
-        http::Method::POST => path_item.post.as_ref(),
-        http::Method::PUT => path_item.put.as_ref(),
-        http::Method::PATCH => path_item.patch.as_ref(),
-        http::Method::DELETE => path_item.delete.as_ref(),
-        http::Method::HEAD => path_item.head.as_ref(),
-        http::Method::OPTIONS => path_item.options.as_ref(),
-        http::Method::TRACE => path_item.trace.as_ref(),
-        _ => None,
-    }
-}
-
-#[cfg(feature = "dashboard")]
-fn infer_route_feature_gates(path: &str) -> Vec<String> {
-    if path.contains("openapi") || path.contains("docs") {
-        vec!["core-openapi".to_string()]
-    } else if path.starts_with("/__rustapi/replays") {
-        vec!["extras-replay".to_string()]
-    } else {
-        Vec::new()
-    }
-}
-
-#[cfg(feature = "dashboard")]
-fn is_dashboard_replay_eligible(path: &str, health_eligible: bool) -> bool {
-    !health_eligible && !path.starts_with("/__rustapi/")
-}
-
-fn add_path_params_to_operation(
-    path: &str,
-    op: &mut rustapi_openapi::Operation,
-    param_schemas: &BTreeMap<String, String>,
-) {
-    let mut params: Vec<String> = Vec::new();
-    let mut in_brace = false;
-    let mut current = String::new();
-
-    for ch in path.chars() {
-        match ch {
-            '{' => {
-                in_brace = true;
-                current.clear();
-            }
-            '}' => {
-                if in_brace {
-                    in_brace = false;
-                    if !current.is_empty() {
-                        params.push(current.clone());
-                    }
-                }
-            }
-            _ => {
-                if in_brace {
-                    current.push(ch);
-                }
-            }
-        }
-    }
-
-    if params.is_empty() {
-        return;
-    }
-
-    let op_params = &mut op.parameters;
-
-    for name in params {
-        let already = op_params
-            .iter()
-            .any(|p| p.location == "path" && p.name == name);
-        if already {
-            continue;
-        }
-
-        // Use custom schema if provided, otherwise infer from name
-        let schema = if let Some(schema_type) = param_schemas.get(&name) {
-            schema_type_to_openapi_schema(schema_type)
-        } else {
-            infer_path_param_schema(&name)
-        };
-
-        op_params.push(rustapi_openapi::Parameter {
-            name,
-            location: "path".to_string(),
-            required: true,
-            description: None,
-            deprecated: None,
-            schema: Some(schema),
-        });
-    }
-}
-
-/// Convert a schema type string to an OpenAPI schema reference
-fn schema_type_to_openapi_schema(schema_type: &str) -> rustapi_openapi::SchemaRef {
-    match schema_type.to_lowercase().as_str() {
-        "uuid" => rustapi_openapi::SchemaRef::Inline(serde_json::json!({
-            "type": "string",
-            "format": "uuid"
-        })),
-        "integer" | "int" | "int64" | "i64" => {
-            rustapi_openapi::SchemaRef::Inline(serde_json::json!({
-                "type": "integer",
-                "format": "int64"
-            }))
-        }
-        "int32" | "i32" => rustapi_openapi::SchemaRef::Inline(serde_json::json!({
-            "type": "integer",
-            "format": "int32"
-        })),
-        "number" | "float" | "f64" | "f32" => {
-            rustapi_openapi::SchemaRef::Inline(serde_json::json!({
-                "type": "number"
-            }))
-        }
-        "boolean" | "bool" => rustapi_openapi::SchemaRef::Inline(serde_json::json!({
-            "type": "boolean"
-        })),
-        _ => rustapi_openapi::SchemaRef::Inline(serde_json::json!({
-            "type": "string"
-        })),
-    }
-}
-
-/// Infer the OpenAPI schema type for a path parameter based on naming conventions.
-///
-/// Common patterns:
-/// - `*_id`, `*Id`, `id` → integer (but NOT *uuid)
-/// - `*_count`, `*_num`, `page`, `limit`, `offset` → integer  
-/// - `*_uuid`, `uuid` → string with uuid format
-/// - `year`, `month`, `day` → integer
-/// - Everything else → string
-fn infer_path_param_schema(name: &str) -> rustapi_openapi::SchemaRef {
-    let lower = name.to_lowercase();
-
-    // UUID patterns (check first to avoid false positive from "id" suffix)
-    let is_uuid = lower == "uuid" || lower.ends_with("_uuid") || lower.ends_with("uuid");
-
-    if is_uuid {
-        return rustapi_openapi::SchemaRef::Inline(serde_json::json!({
-            "type": "string",
-            "format": "uuid"
-        }));
-    }
-
-    // Integer patterns
-    // Integer patterns
-    let is_integer = lower == "page"
-        || lower == "limit"
-        || lower == "offset"
-        || lower == "count"
-        || lower.ends_with("_count")
-        || lower.ends_with("_num")
-        || lower == "year"
-        || lower == "month"
-        || lower == "day"
-        || lower == "index"
-        || lower == "position";
-
-    if is_integer {
-        rustapi_openapi::SchemaRef::Inline(serde_json::json!({
-            "type": "integer",
-            "format": "int64"
-        }))
-    } else {
-        rustapi_openapi::SchemaRef::Inline(serde_json::json!({ "type": "string" }))
-    }
-}
-
-/// Normalize a prefix for OpenAPI paths.
-///
-/// Ensures the prefix:
-/// - Starts with exactly one leading slash
-/// - Has no trailing slash (unless it's just "/")
-/// - Has no double slashes
-fn normalize_prefix_for_openapi(prefix: &str) -> String {
-    // Handle empty string
-    if prefix.is_empty() {
-        return "/".to_string();
-    }
-
-    // Split by slashes and filter out empty segments (handles multiple slashes)
-    let segments: Vec<&str> = prefix.split('/').filter(|s| !s.is_empty()).collect();
-
-    // If no segments after filtering, return root
-    if segments.is_empty() {
-        return "/".to_string();
-    }
-
-    // Build the normalized prefix with leading slash
-    let mut result = String::with_capacity(prefix.len() + 1);
-    for segment in segments {
-        result.push('/');
-        result.push_str(segment);
-    }
-
-    result
-}
-
 impl Default for RustApi {
     fn default() -> Self {
         Self::new()
-    }
-}
-
-/// Check Basic Auth header against expected credentials
-#[cfg(feature = "swagger-ui")]
-fn check_basic_auth(req: &crate::Request, expected: &str) -> bool {
-    req.headers()
-        .get(http::header::AUTHORIZATION)
-        .and_then(|v| v.to_str().ok())
-        .map(|auth| auth == expected)
-        .unwrap_or(false)
-}
-
-/// Create 401 Unauthorized response with WWW-Authenticate header
-#[cfg(feature = "swagger-ui")]
-fn unauthorized_response() -> crate::Response {
-    http::Response::builder()
-        .status(http::StatusCode::UNAUTHORIZED)
-        .header(
-            http::header::WWW_AUTHENTICATE,
-            "Basic realm=\"API Documentation\"",
-        )
-        .header(http::header::CONTENT_TYPE, "text/plain")
-        .body(crate::response::Body::from("Unauthorized"))
-        .unwrap()
-}
-
-/// Configuration builder for RustAPI with auto-routes
-pub struct RustApiConfig {
-    docs_path: Option<String>,
-    docs_enabled: bool,
-    api_title: String,
-    api_version: String,
-    api_description: Option<String>,
-    body_limit: Option<usize>,
-    layers: LayerStack,
-}
-
-impl Default for RustApiConfig {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-impl RustApiConfig {
-    pub fn new() -> Self {
-        Self {
-            docs_path: Some("/docs".to_string()),
-            docs_enabled: true,
-            api_title: "RustAPI".to_string(),
-            api_version: "1.0.0".to_string(),
-            api_description: None,
-            body_limit: None,
-            layers: LayerStack::new(),
-        }
-    }
-
-    /// Set the docs path (default: "/docs")
-    pub fn docs_path(mut self, path: impl Into<String>) -> Self {
-        self.docs_path = Some(path.into());
-        self
-    }
-
-    /// Enable or disable docs (default: true)
-    pub fn docs_enabled(mut self, enabled: bool) -> Self {
-        self.docs_enabled = enabled;
-        self
-    }
-
-    /// Set OpenAPI info
-    pub fn openapi_info(
-        mut self,
-        title: impl Into<String>,
-        version: impl Into<String>,
-        description: Option<impl Into<String>>,
-    ) -> Self {
-        self.api_title = title.into();
-        self.api_version = version.into();
-        self.api_description = description.map(|d| d.into());
-        self
-    }
-
-    /// Set body size limit
-    pub fn body_limit(mut self, limit: usize) -> Self {
-        self.body_limit = Some(limit);
-        self
-    }
-
-    /// Add a middleware layer
-    pub fn layer<L>(mut self, layer: L) -> Self
-    where
-        L: MiddlewareLayer,
-    {
-        self.layers.push(Box::new(layer));
-        self
-    }
-
-    /// Build the RustApi instance
-    pub fn build(self) -> RustApi {
-        let mut app = RustApi::new().mount_auto_routes_grouped();
-
-        // Apply configuration
-        if let Some(limit) = self.body_limit {
-            app = app.body_limit(limit);
-        }
-
-        app = app.openapi_info(
-            &self.api_title,
-            &self.api_version,
-            self.api_description.as_deref(),
-        );
-
-        #[cfg(feature = "swagger-ui")]
-        if self.docs_enabled {
-            if let Some(path) = self.docs_path {
-                app = app.docs(&path);
-            }
-        }
-
-        // Apply layers
-        // Note: layers are applied in reverse order in RustApi::layer logic (pushing to vec)
-        app.layers.extend(self.layers);
-
-        app
-    }
-
-    /// Build and run the server
-    pub async fn run(
-        self,
-        addr: impl AsRef<str>,
-    ) -> Result<(), Box<dyn std::error::Error + Send + Sync>> {
-        self.build().run(addr.as_ref()).await
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::RustApi;
-    use crate::extract::{FromRequestParts, State};
-    use crate::path_params::PathParams;
-    use crate::request::Request;
-    use crate::router::{get, post, Router};
-    use bytes::Bytes;
-    use http::Method;
-    use proptest::prelude::*;
-
-    #[test]
-    fn state_is_available_via_extractor() {
-        let app = RustApi::new().state(123u32);
-        let router = app.into_router();
-
-        let req = http::Request::builder()
-            .method(Method::GET)
-            .uri("/test")
-            .body(())
-            .unwrap();
-        let (parts, _) = req.into_parts();
-
-        let request = Request::new(
-            parts,
-            crate::request::BodyVariant::Buffered(Bytes::new()),
-            router.state_ref(),
-            PathParams::new(),
-        );
-        let State(value) = State::<u32>::from_request_parts(&request).unwrap();
-        assert_eq!(value, 123u32);
-    }
-
-    #[test]
-    fn test_path_param_type_inference_integer() {
-        use super::infer_path_param_schema;
-
-        // Test common integer patterns
-        let int_params = [
-            "page",
-            "limit",
-            "offset",
-            "count",
-            "item_count",
-            "year",
-            "month",
-            "day",
-            "index",
-            "position",
-        ];
-
-        for name in int_params {
-            let schema = infer_path_param_schema(name);
-            match schema {
-                rustapi_openapi::SchemaRef::Inline(v) => {
-                    assert_eq!(
-                        v.get("type").and_then(|v| v.as_str()),
-                        Some("integer"),
-                        "Expected '{}' to be inferred as integer",
-                        name
-                    );
-                }
-                _ => panic!("Expected inline schema for '{}'", name),
-            }
-        }
-    }
-
-    #[test]
-    fn test_path_param_type_inference_uuid() {
-        use super::infer_path_param_schema;
-
-        // Test UUID patterns
-        let uuid_params = ["uuid", "user_uuid", "sessionUuid"];
-
-        for name in uuid_params {
-            let schema = infer_path_param_schema(name);
-            match schema {
-                rustapi_openapi::SchemaRef::Inline(v) => {
-                    assert_eq!(
-                        v.get("type").and_then(|v| v.as_str()),
-                        Some("string"),
-                        "Expected '{}' to be inferred as string",
-                        name
-                    );
-                    assert_eq!(
-                        v.get("format").and_then(|v| v.as_str()),
-                        Some("uuid"),
-                        "Expected '{}' to have uuid format",
-                        name
-                    );
-                }
-                _ => panic!("Expected inline schema for '{}'", name),
-            }
-        }
-    }
-
-    #[test]
-    fn test_path_param_type_inference_string() {
-        use super::infer_path_param_schema;
-
-        // Test string (default) patterns
-        let string_params = [
-            "name", "slug", "code", "token", "username", "id", "user_id", "userId", "postId",
-        ];
-
-        for name in string_params {
-            let schema = infer_path_param_schema(name);
-            match schema {
-                rustapi_openapi::SchemaRef::Inline(v) => {
-                    assert_eq!(
-                        v.get("type").and_then(|v| v.as_str()),
-                        Some("string"),
-                        "Expected '{}' to be inferred as string",
-                        name
-                    );
-                    assert!(
-                        v.get("format").is_none()
-                            || v.get("format").and_then(|v| v.as_str()) != Some("uuid"),
-                        "Expected '{}' to NOT have uuid format",
-                        name
-                    );
-                }
-                _ => panic!("Expected inline schema for '{}'", name),
-            }
-        }
-    }
-
-    #[test]
-    fn test_schema_type_to_openapi_schema() {
-        use super::schema_type_to_openapi_schema;
-
-        // Test UUID schema
-        let uuid_schema = schema_type_to_openapi_schema("uuid");
-        match uuid_schema {
-            rustapi_openapi::SchemaRef::Inline(v) => {
-                assert_eq!(v.get("type").and_then(|v| v.as_str()), Some("string"));
-                assert_eq!(v.get("format").and_then(|v| v.as_str()), Some("uuid"));
-            }
-            _ => panic!("Expected inline schema for uuid"),
-        }
-
-        // Test integer schemas
-        for schema_type in ["integer", "int", "int64", "i64"] {
-            let schema = schema_type_to_openapi_schema(schema_type);
-            match schema {
-                rustapi_openapi::SchemaRef::Inline(v) => {
-                    assert_eq!(v.get("type").and_then(|v| v.as_str()), Some("integer"));
-                    assert_eq!(v.get("format").and_then(|v| v.as_str()), Some("int64"));
-                }
-                _ => panic!("Expected inline schema for {}", schema_type),
-            }
-        }
-
-        // Test int32 schema
-        let int32_schema = schema_type_to_openapi_schema("int32");
-        match int32_schema {
-            rustapi_openapi::SchemaRef::Inline(v) => {
-                assert_eq!(v.get("type").and_then(|v| v.as_str()), Some("integer"));
-                assert_eq!(v.get("format").and_then(|v| v.as_str()), Some("int32"));
-            }
-            _ => panic!("Expected inline schema for int32"),
-        }
-
-        // Test number/float schema
-        for schema_type in ["number", "float"] {
-            let schema = schema_type_to_openapi_schema(schema_type);
-            match schema {
-                rustapi_openapi::SchemaRef::Inline(v) => {
-                    assert_eq!(v.get("type").and_then(|v| v.as_str()), Some("number"));
-                }
-                _ => panic!("Expected inline schema for {}", schema_type),
-            }
-        }
-
-        // Test boolean schema
-        for schema_type in ["boolean", "bool"] {
-            let schema = schema_type_to_openapi_schema(schema_type);
-            match schema {
-                rustapi_openapi::SchemaRef::Inline(v) => {
-                    assert_eq!(v.get("type").and_then(|v| v.as_str()), Some("boolean"));
-                }
-                _ => panic!("Expected inline schema for {}", schema_type),
-            }
-        }
-
-        // Test string schema (default)
-        let string_schema = schema_type_to_openapi_schema("string");
-        match string_schema {
-            rustapi_openapi::SchemaRef::Inline(v) => {
-                assert_eq!(v.get("type").and_then(|v| v.as_str()), Some("string"));
-            }
-            _ => panic!("Expected inline schema for string"),
-        }
-    }
-
-    // **Feature: router-nesting, Property 11: OpenAPI Integration**
-    //
-    // For any nested routes with OpenAPI operations, the operations should appear
-    // in the parent's OpenAPI spec with prefixed paths and preserved metadata.
-    //
-    // **Validates: Requirements 4.1, 4.2**
-    proptest! {
-        #![proptest_config(ProptestConfig::with_cases(100))]
-
-        /// Property: Nested routes appear in OpenAPI spec with prefixed paths
-        ///
-        /// For any router with routes nested under a prefix, all routes should
-        /// appear in the OpenAPI spec with the prefix prepended to their paths.
-        #[test]
-        fn prop_nested_routes_in_openapi_spec(
-            // Generate prefix segments
-            prefix_segments in prop::collection::vec("[a-z][a-z0-9]{0,5}", 1..3),
-            // Generate route path segments
-            route_segments in prop::collection::vec("[a-z][a-z0-9]{0,5}", 1..3),
-            has_param in any::<bool>(),
-        ) {
-            async fn handler() -> &'static str { "handler" }
-
-            // Build the prefix
-            let prefix = format!("/{}", prefix_segments.join("/"));
-
-            // Build the route path
-            let mut route_path = format!("/{}", route_segments.join("/"));
-            if has_param {
-                route_path.push_str("/{id}");
-            }
-
-            // Create nested router and nest it through RustApi
-            let nested_router = Router::new().route(&route_path, get(handler));
-            let app = RustApi::new().nest(&prefix, nested_router);
-
-            // Build expected prefixed path for OpenAPI (uses {param} format)
-            let expected_openapi_path = format!("{}{}", prefix, route_path);
-
-            // Get the OpenAPI spec
-            let spec = app.openapi_spec();
-
-            // Property: The prefixed route should exist in OpenAPI paths
-            prop_assert!(
-                spec.paths.contains_key(&expected_openapi_path),
-                "Expected OpenAPI path '{}' not found. Available paths: {:?}",
-                expected_openapi_path,
-                spec.paths.keys().collect::<Vec<_>>()
-            );
-
-            // Property: The path item should have a GET operation
-            let path_item = spec.paths.get(&expected_openapi_path).unwrap();
-            prop_assert!(
-                path_item.get.is_some(),
-                "GET operation should exist for path '{}'",
-                expected_openapi_path
-            );
-        }
-
-        /// Property: Multiple HTTP methods are preserved in OpenAPI spec after nesting
-        ///
-        /// For any router with routes having multiple HTTP methods, nesting should
-        /// preserve all method operations in the OpenAPI spec.
-        #[test]
-        fn prop_multiple_methods_preserved_in_openapi(
-            prefix_segments in prop::collection::vec("[a-z][a-z0-9]{0,5}", 1..3),
-            route_segments in prop::collection::vec("[a-z][a-z0-9]{0,5}", 1..3),
-        ) {
-            async fn get_handler() -> &'static str { "get" }
-            async fn post_handler() -> &'static str { "post" }
-
-            // Build the prefix and route path
-            let prefix = format!("/{}", prefix_segments.join("/"));
-            let route_path = format!("/{}", route_segments.join("/"));
-
-            // Create nested router with both GET and POST using separate routes
-            // Since MethodRouter doesn't have chaining methods, we create two routes
-            let get_route_path = format!("{}/get", route_path);
-            let post_route_path = format!("{}/post", route_path);
-            let nested_router = Router::new()
-                .route(&get_route_path, get(get_handler))
-                .route(&post_route_path, post(post_handler));
-            let app = RustApi::new().nest(&prefix, nested_router);
-
-            // Build expected prefixed paths for OpenAPI
-            let expected_get_path = format!("{}{}", prefix, get_route_path);
-            let expected_post_path = format!("{}{}", prefix, post_route_path);
-
-            // Get the OpenAPI spec
-            let spec = app.openapi_spec();
-
-            // Property: Both paths should exist
-            prop_assert!(
-                spec.paths.contains_key(&expected_get_path),
-                "Expected OpenAPI path '{}' not found",
-                expected_get_path
-            );
-            prop_assert!(
-                spec.paths.contains_key(&expected_post_path),
-                "Expected OpenAPI path '{}' not found",
-                expected_post_path
-            );
-
-            // Property: GET operation should exist on get path
-            let get_path_item = spec.paths.get(&expected_get_path).unwrap();
-            prop_assert!(
-                get_path_item.get.is_some(),
-                "GET operation should exist for path '{}'",
-                expected_get_path
-            );
-
-            // Property: POST operation should exist on post path
-            let post_path_item = spec.paths.get(&expected_post_path).unwrap();
-            prop_assert!(
-                post_path_item.post.is_some(),
-                "POST operation should exist for path '{}'",
-                expected_post_path
-            );
-        }
-
-        /// Property: Path parameters are added to OpenAPI operations after nesting
-        ///
-        /// For any nested route with path parameters, the OpenAPI operation should
-        /// include the path parameters.
-        #[test]
-        fn prop_path_params_in_openapi_after_nesting(
-            prefix_segments in prop::collection::vec("[a-z][a-z0-9]{0,5}", 1..3),
-            param_name in "[a-z][a-z0-9]{0,5}",
-        ) {
-            async fn handler() -> &'static str { "handler" }
-
-            // Build the prefix and route path with parameter
-            let prefix = format!("/{}", prefix_segments.join("/"));
-            let route_path = format!("/{{{}}}", param_name);
-
-            // Create nested router
-            let nested_router = Router::new().route(&route_path, get(handler));
-            let app = RustApi::new().nest(&prefix, nested_router);
-
-            // Build expected prefixed path for OpenAPI
-            let expected_openapi_path = format!("{}{}", prefix, route_path);
-
-            // Get the OpenAPI spec
-            let spec = app.openapi_spec();
-
-            // Property: The path should exist
-            prop_assert!(
-                spec.paths.contains_key(&expected_openapi_path),
-                "Expected OpenAPI path '{}' not found",
-                expected_openapi_path
-            );
-
-            // Property: The GET operation should have the path parameter
-            let path_item = spec.paths.get(&expected_openapi_path).unwrap();
-            let get_op = path_item.get.as_ref().unwrap();
-
-            prop_assert!(
-                !get_op.parameters.is_empty(),
-                "Operation should have parameters for path '{}'",
-                expected_openapi_path
-            );
-
-            let params = &get_op.parameters;
-            let has_param = params.iter().any(|p| p.name == param_name && p.location == "path");
-            prop_assert!(
-                has_param,
-                "Path parameter '{}' should exist in operation parameters. Found: {:?}",
-                param_name,
-                params.iter().map(|p| &p.name).collect::<Vec<_>>()
-            );
-        }
-    }
-
-    // **Feature: router-nesting, Property 13: RustApi Integration**
-    //
-    // For any router nested through `RustApi::new().nest()`, the behavior should be
-    // identical to nesting through `Router::new().nest()`, and routes should appear
-    // in the OpenAPI spec.
-    //
-    // **Validates: Requirements 6.1, 6.2**
-    proptest! {
-        #![proptest_config(ProptestConfig::with_cases(100))]
-
-        /// Property: RustApi::nest delegates to Router::nest and produces identical route registration
-        ///
-        /// For any router with routes nested under a prefix, nesting through RustApi
-        /// should produce the same route registration as nesting through Router directly.
-        #[test]
-        fn prop_rustapi_nest_delegates_to_router_nest(
-            prefix_segments in prop::collection::vec("[a-z][a-z0-9]{0,5}", 1..3),
-            route_segments in prop::collection::vec("[a-z][a-z0-9]{0,5}", 1..3),
-            has_param in any::<bool>(),
-        ) {
-            async fn handler() -> &'static str { "handler" }
-
-            // Build the prefix
-            let prefix = format!("/{}", prefix_segments.join("/"));
-
-            // Build the route path
-            let mut route_path = format!("/{}", route_segments.join("/"));
-            if has_param {
-                route_path.push_str("/{id}");
-            }
-
-            // Create nested router
-            let nested_router_for_rustapi = Router::new().route(&route_path, get(handler));
-            let nested_router_for_router = Router::new().route(&route_path, get(handler));
-
-            // Nest through RustApi
-            let rustapi_app = RustApi::new().nest(&prefix, nested_router_for_rustapi);
-            let rustapi_router = rustapi_app.into_router();
-
-            // Nest through Router directly
-            let router_app = Router::new().nest(&prefix, nested_router_for_router);
-
-            // Property: Both should have the same registered routes
-            let rustapi_routes = rustapi_router.registered_routes();
-            let router_routes = router_app.registered_routes();
-
-            prop_assert_eq!(
-                rustapi_routes.len(),
-                router_routes.len(),
-                "RustApi and Router should have same number of routes"
-            );
-
-            // Property: All routes from Router should exist in RustApi
-            for (path, info) in router_routes {
-                prop_assert!(
-                    rustapi_routes.contains_key(path),
-                    "Route '{}' from Router should exist in RustApi routes",
-                    path
-                );
-
-                let rustapi_info = rustapi_routes.get(path).unwrap();
-                prop_assert_eq!(
-                    &info.path, &rustapi_info.path,
-                    "Display paths should match for route '{}'",
-                    path
-                );
-                prop_assert_eq!(
-                    info.methods.len(), rustapi_info.methods.len(),
-                    "Method count should match for route '{}'",
-                    path
-                );
-            }
-        }
-
-        /// Property: RustApi::nest includes nested routes in OpenAPI spec
-        ///
-        /// For any router with routes nested through RustApi, all routes should
-        /// appear in the OpenAPI specification with prefixed paths.
-        #[test]
-        fn prop_rustapi_nest_includes_routes_in_openapi(
-            prefix_segments in prop::collection::vec("[a-z][a-z0-9]{0,5}", 1..3),
-            route_segments in prop::collection::vec("[a-z][a-z0-9]{0,5}", 1..3),
-            has_param in any::<bool>(),
-        ) {
-            async fn handler() -> &'static str { "handler" }
-
-            // Build the prefix
-            let prefix = format!("/{}", prefix_segments.join("/"));
-
-            // Build the route path
-            let mut route_path = format!("/{}", route_segments.join("/"));
-            if has_param {
-                route_path.push_str("/{id}");
-            }
-
-            // Create nested router and nest through RustApi
-            let nested_router = Router::new().route(&route_path, get(handler));
-            let app = RustApi::new().nest(&prefix, nested_router);
-
-            // Build expected prefixed path for OpenAPI
-            let expected_openapi_path = format!("{}{}", prefix, route_path);
-
-            // Get the OpenAPI spec
-            let spec = app.openapi_spec();
-
-            // Property: The prefixed route should exist in OpenAPI paths
-            prop_assert!(
-                spec.paths.contains_key(&expected_openapi_path),
-                "Expected OpenAPI path '{}' not found. Available paths: {:?}",
-                expected_openapi_path,
-                spec.paths.keys().collect::<Vec<_>>()
-            );
-
-            // Property: The path item should have a GET operation
-            let path_item = spec.paths.get(&expected_openapi_path).unwrap();
-            prop_assert!(
-                path_item.get.is_some(),
-                "GET operation should exist for path '{}'",
-                expected_openapi_path
-            );
-        }
-
-        /// Property: RustApi::nest route matching is identical to Router::nest
-        ///
-        /// For any nested route, matching through RustApi should produce the same
-        /// result as matching through Router directly.
-        #[test]
-        fn prop_rustapi_nest_route_matching_identical(
-            prefix_segments in prop::collection::vec("[a-z][a-z0-9]{1,5}", 1..2),
-            route_segments in prop::collection::vec("[a-z][a-z0-9]{1,5}", 1..2),
-            param_value in "[a-z0-9]{1,10}",
-        ) {
-            use crate::router::RouteMatch;
-
-            async fn handler() -> &'static str { "handler" }
-
-            // Build the prefix and route path with parameter
-            let prefix = format!("/{}", prefix_segments.join("/"));
-            let route_path = format!("/{}/{{id}}", route_segments.join("/"));
-
-            // Create nested routers
-            let nested_router_for_rustapi = Router::new().route(&route_path, get(handler));
-            let nested_router_for_router = Router::new().route(&route_path, get(handler));
-
-            // Nest through both RustApi and Router
-            let rustapi_app = RustApi::new().nest(&prefix, nested_router_for_rustapi);
-            let rustapi_router = rustapi_app.into_router();
-            let router_app = Router::new().nest(&prefix, nested_router_for_router);
-
-            // Build the full path to match
-            let full_path = format!("{}/{}/{}", prefix, route_segments.join("/"), param_value);
-
-            // Match through both
-            let rustapi_match = rustapi_router.match_route(&full_path, &Method::GET);
-            let router_match = router_app.match_route(&full_path, &Method::GET);
-
-            // Property: Both should return Found with same parameters
-            match (rustapi_match, router_match) {
-                (RouteMatch::Found { params: rustapi_params, .. }, RouteMatch::Found { params: router_params, .. }) => {
-                    prop_assert_eq!(
-                        rustapi_params.len(),
-                        router_params.len(),
-                        "Parameter count should match"
-                    );
-                    for (key, value) in &router_params {
-                        prop_assert!(
-                            rustapi_params.contains_key(key),
-                            "RustApi should have parameter '{}'",
-                            key
-                        );
-                        prop_assert_eq!(
-                            rustapi_params.get(key).unwrap(),
-                            value,
-                            "Parameter '{}' value should match",
-                            key
-                        );
-                    }
-                }
-                (rustapi_result, router_result) => {
-                    prop_assert!(
-                        false,
-                        "Both should return Found, but RustApi returned {:?} and Router returned {:?}",
-                        match rustapi_result {
-                            RouteMatch::Found { .. } => "Found",
-                            RouteMatch::NotFound => "NotFound",
-                            RouteMatch::MethodNotAllowed { .. } => "MethodNotAllowed",
-                        },
-                        match router_result {
-                            RouteMatch::Found { .. } => "Found",
-                            RouteMatch::NotFound => "NotFound",
-                            RouteMatch::MethodNotAllowed { .. } => "MethodNotAllowed",
-                        }
-                    );
-                }
-            }
-        }
-    }
-
-    /// Unit test: Verify OpenAPI operations are propagated during nesting
-    #[test]
-    fn test_openapi_operations_propagated_during_nesting() {
-        async fn list_users() -> &'static str {
-            "list users"
-        }
-        async fn get_user() -> &'static str {
-            "get user"
-        }
-        async fn create_user() -> &'static str {
-            "create user"
-        }
-
-        // Create nested router with multiple routes
-        // Note: We use separate routes since MethodRouter doesn't support chaining
-        let users_router = Router::new()
-            .route("/", get(list_users))
-            .route("/create", post(create_user))
-            .route("/{id}", get(get_user));
-
-        // Nest under /api/v1/users
-        let app = RustApi::new().nest("/api/v1/users", users_router);
-
-        let spec = app.openapi_spec();
-
-        // Verify /api/v1/users path exists with GET
-        assert!(
-            spec.paths.contains_key("/api/v1/users"),
-            "Should have /api/v1/users path"
-        );
-        let users_path = spec.paths.get("/api/v1/users").unwrap();
-        assert!(users_path.get.is_some(), "Should have GET operation");
-
-        // Verify /api/v1/users/create path exists with POST
-        assert!(
-            spec.paths.contains_key("/api/v1/users/create"),
-            "Should have /api/v1/users/create path"
-        );
-        let create_path = spec.paths.get("/api/v1/users/create").unwrap();
-        assert!(create_path.post.is_some(), "Should have POST operation");
-
-        // Verify /api/v1/users/{id} path exists with GET
-        assert!(
-            spec.paths.contains_key("/api/v1/users/{id}"),
-            "Should have /api/v1/users/{{id}} path"
-        );
-        let user_path = spec.paths.get("/api/v1/users/{id}").unwrap();
-        assert!(
-            user_path.get.is_some(),
-            "Should have GET operation for user by id"
-        );
-
-        // Verify path parameter is added
-        let get_user_op = user_path.get.as_ref().unwrap();
-        assert!(!get_user_op.parameters.is_empty(), "Should have parameters");
-        let params = &get_user_op.parameters;
-        assert!(
-            params
-                .iter()
-                .any(|p| p.name == "id" && p.location == "path"),
-            "Should have 'id' path parameter"
-        );
-    }
-
-    /// Unit test: Verify nested routes don't appear without nesting
-    #[test]
-    fn test_openapi_spec_empty_without_routes() {
-        let app = RustApi::new();
-        let spec = app.openapi_spec();
-
-        // Should have no paths (except potentially default ones)
-        assert!(
-            spec.paths.is_empty(),
-            "OpenAPI spec should have no paths without routes"
-        );
-    }
-
-    /// Unit test: Verify RustApi::nest delegates correctly to Router::nest
-    ///
-    /// **Feature: router-nesting, Property 13: RustApi Integration**
-    /// **Validates: Requirements 6.1, 6.2**
-    #[test]
-    fn test_rustapi_nest_delegates_to_router_nest() {
-        use crate::router::RouteMatch;
-
-        async fn list_users() -> &'static str {
-            "list users"
-        }
-        async fn get_user() -> &'static str {
-            "get user"
-        }
-        async fn create_user() -> &'static str {
-            "create user"
-        }
-
-        // Create nested router with multiple routes
-        let users_router = Router::new()
-            .route("/", get(list_users))
-            .route("/create", post(create_user))
-            .route("/{id}", get(get_user));
-
-        // Nest through RustApi
-        let app = RustApi::new().nest("/api/v1/users", users_router);
-        let router = app.into_router();
-
-        // Verify routes are registered correctly
-        let routes = router.registered_routes();
-        assert_eq!(routes.len(), 3, "Should have 3 routes registered");
-
-        // Verify route paths
-        assert!(
-            routes.contains_key("/api/v1/users"),
-            "Should have /api/v1/users route"
-        );
-        assert!(
-            routes.contains_key("/api/v1/users/create"),
-            "Should have /api/v1/users/create route"
-        );
-        assert!(
-            routes.contains_key("/api/v1/users/:id"),
-            "Should have /api/v1/users/:id route"
-        );
-
-        // Verify route matching works
-        match router.match_route("/api/v1/users", &Method::GET) {
-            RouteMatch::Found { params, .. } => {
-                assert!(params.is_empty(), "Root route should have no params");
-            }
-            _ => panic!("GET /api/v1/users should be found"),
-        }
-
-        match router.match_route("/api/v1/users/create", &Method::POST) {
-            RouteMatch::Found { params, .. } => {
-                assert!(params.is_empty(), "Create route should have no params");
-            }
-            _ => panic!("POST /api/v1/users/create should be found"),
-        }
-
-        match router.match_route("/api/v1/users/123", &Method::GET) {
-            RouteMatch::Found { params, .. } => {
-                assert_eq!(
-                    params.get("id"),
-                    Some(&"123".to_string()),
-                    "Should extract id param"
-                );
-            }
-            _ => panic!("GET /api/v1/users/123 should be found"),
-        }
-
-        // Verify method not allowed
-        match router.match_route("/api/v1/users", &Method::DELETE) {
-            RouteMatch::MethodNotAllowed { allowed } => {
-                assert!(allowed.contains(&Method::GET), "Should allow GET");
-            }
-            _ => panic!("DELETE /api/v1/users should return MethodNotAllowed"),
-        }
-    }
-
-    /// Unit test: Verify RustApi::nest includes routes in OpenAPI spec
-    ///
-    /// **Feature: router-nesting, Property 13: RustApi Integration**
-    /// **Validates: Requirements 6.1, 6.2**
-    #[test]
-    fn test_rustapi_nest_includes_routes_in_openapi_spec() {
-        async fn list_items() -> &'static str {
-            "list items"
-        }
-        async fn get_item() -> &'static str {
-            "get item"
-        }
-
-        // Create nested router
-        let items_router = Router::new()
-            .route("/", get(list_items))
-            .route("/{item_id}", get(get_item));
-
-        // Nest through RustApi
-        let app = RustApi::new().nest("/api/items", items_router);
-
-        // Verify OpenAPI spec
-        let spec = app.openapi_spec();
-
-        // Verify paths exist
-        assert!(
-            spec.paths.contains_key("/api/items"),
-            "Should have /api/items in OpenAPI"
-        );
-        assert!(
-            spec.paths.contains_key("/api/items/{item_id}"),
-            "Should have /api/items/{{item_id}} in OpenAPI"
-        );
-
-        // Verify operations
-        let list_path = spec.paths.get("/api/items").unwrap();
-        assert!(
-            list_path.get.is_some(),
-            "Should have GET operation for /api/items"
-        );
-
-        let get_path = spec.paths.get("/api/items/{item_id}").unwrap();
-        assert!(
-            get_path.get.is_some(),
-            "Should have GET operation for /api/items/{{item_id}}"
-        );
-
-        // Verify path parameter is added
-        let get_op = get_path.get.as_ref().unwrap();
-        assert!(!get_op.parameters.is_empty(), "Should have parameters");
-        let params = &get_op.parameters;
-        assert!(
-            params
-                .iter()
-                .any(|p| p.name == "item_id" && p.location == "path"),
-            "Should have 'item_id' path parameter"
-        );
     }
 }
