+++ crates/rustapi-core/src/app/mod.rs
@@ -1,0 +1,16 @@
+﻿//! RustApi application builder
+
+mod builder;
+mod config;
+mod dispatcher;
+mod helpers;
+mod production;
+mod types;
+
+#[cfg(test)]
+mod tests;
+
+pub use config::RustApiConfig;
+pub use dispatcher::RequestDispatcher;
+pub use production::ProductionDefaultsConfig;
+pub use types::RustApi;
